refactor(web): Remove unused code and rename files for consistency
