--- apps/web/src/hooks/useTheme.ts
@@ -1,10 +1,0 @@
-import { ThemeProviderContext } from "@/providers/theme-provider"
-import { useContext } from "react"
-
-export const useTheme = () => {
-    const context = useContext(ThemeProviderContext)
-
-    if (context === undefined) throw new Error("useTheme must be used within a ThemeProvider")
-
-    return context
-}
