+++ apps/web/src/main.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import ReactDOM from "react-dom/client"
 
-import { ThemeProvider } from "@/providers/theme-provider.tsx"
+import { ThemeProvider } from "@/providers/theme.tsx"
 import App from "./app.tsx"
 
 import "./global.css"
